Add temperature averaging to solver and Q2D fd case
The solver has new code implementing temperature
averaging in fullyDeveloped.H, being this file called
in the main Q2DmhdFoam.C. Some auxiliary fields are
initialized in createFields.
Furthermore, the new boolean variable 'fd' is defined
in transportProperties so the new is called or not,
so this functionality can be easily taken out the
same way Ubar=0 takes out the pump.
The case Q2DfullyDeveloped implements now this