Update solver with improved pump and writing files
Files are modified to use the gradP in the UEqn
momentum equation and in a 3D way, i.e. using
the flowDirection variable as in channelFoam
to automatically detect the direction of Ubar
defined by the user.

Also, now ioFiles.H and other lines of code
define the writing procedure of some variables
to output.out file each simulation. These
variables are related to gradP calculation and
the max residual evolution for reaching steady
state.