Add new validation directory for cases Gr=0
This is made because we had problems with Grashoff
and velocity profiles worked weirdly and we want
to check if the problem is in heat deposition
generation. So validation/tepot_nullGr will
test the goodness of Q2DmhdFoam without thermal
deposition and with varying Hartman numbers.
We fix Re at 500, on the limit between Q2D
turbulence and laminar according to Smolentsev

This commit includes two samples from tepot,
still missing a third one with Ha=2000