Update in the Poiseuille's case
g was 9.8 while this case should have
no gravity. It's now changed to 0 in
the baseCase files. However it doesn't
affect the results

There are now some changes in the
convergence criteria, not too relevant

The validation.py script has some
changes to improve the program and
uses now only one cell in the direction
of the flow, but the result is the same.
The most important change is to multiply
the deltaP extracted from OpenFOAM by
the density of the fluid, so now the order
of magnitude is correct